WIP: split template loading into multiple files
